requirements.txt
accelerate==1.3.0
deepspeed==0.16.3
icecream==2.1.4
numpy==2.2.3
opencv_python==4.11.0.86
packaging==24.2
peft==0.14.0
Pillow==11.1.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
nltk==3.9.1
wandb==0.19.8